TODO(spoon): warn on long values passing between Java and JavaScript

Patch by: spoon (LongLib/LongLibTest), me (most everything else)
Review by: fabbott (LongLib/JRE changes), bobv & spoon (TBR, everything else)



git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@2145 8db76d5a-ed1c-0410-87a9-c151d255dfc7

+++ user/src/com/google/gwt/user/client/Cookies.java
@@ -147,7 +147,7 @@ public class Cookies {
   }-*/;
 
   private static native void setCookieImpl(String name, String value,
-      long expires, String domain, String path, boolean secure) /*-{
+      double expires, String domain, String path, boolean secure) /*-{
     var c = encodeURIComponent(name) + '=' + encodeURIComponent(value);
     if ( expires )
       c += ';expires=' + (new Date(expires)).toGMTString();
